+++ Modules/Intent.Modules.WindowsServiceHost/Templates/AppStartupFile.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿#nullable enable
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using Intent.Modules.Common.CSharp.AppStartup;
@@ -26,6 +27,7 @@ internal class AppStartupFile : IAppStartupFile
     private readonly List<ContainerRegistrationRequest> _containerRegistrationRequests = new();
     private readonly List<ServiceConfigurationRequest> _serviceConfigurationRequests = new();
     private readonly List<ApplicationBuilderRegistrationRequest> _applicationBuilderRegistrationRequests = new();
+    private IHasCSharpStatements? _configureServicesBlock;
 
     public AppStartupFile(CSharpTemplateBase<object> template)
     {
@@ -41,7 +43,7 @@ internal class AppStartupFile : IAppStartupFile
         if (template is ProgramTemplate)
         {
             var netSettings = template.OutputTarget.GetProject()?.InternalElement?.AsCSharpProjectNETModel()?.GetNETSettings();
-            _usesMinimalHostingModel = true;//netSettings?.UseMinimalHostingModel() == true;
+            _usesMinimalHostingModel = true;
             _usesTopLevelStatements = netSettings?.UseTopLevelStatements() == true;
         }
 
@@ -129,47 +131,6 @@ internal class AppStartupFile : IAppStartupFile
                     }
                 });
 
-                /*
-                _applicationBuilderRegistrationRequests.Add(ApplicationBuilderRegistrationRequest.ToRegister("UseHttpsRedirection").WithPriority(-30));
-                _applicationBuilderRegistrationRequests.Add(ApplicationBuilderRegistrationRequest.ToRegister("UseRouting").WithPriority(-20));
-                _applicationBuilderRegistrationRequests.Add(ApplicationBuilderRegistrationRequest.ToRegister("UseAuthorization").WithPriority(-5));
-                */
-                /*
-                // To match previous behaviour (and avoid changes for upgrading clients) where would always be generated
-                AddAppConfigurationLambdaInternal(
-                    "UseEndpoints",
-                    new[] { "endpoints" },
-                    (_, lambdaStatement, _) =>
-                    {
-                        // Would previously create an empty line in the block, so we replicate that, but remove the empty line
-                        // if anything else gets added.
-                        var cSharpStatement = new CSharpStatement(string.Empty).AddMetadata("is-startup-file-placeholder", true);
-                        lambdaStatement.AddStatement(cSharpStatement);
-
-                        _cSharpFile.AfterBuild(_ =>
-                        {
-                            if (lambdaStatement.Statements.Any(x => !Equals(x, cSharpStatement)))
-                            {
-                                cSharpStatement.Remove();
-                            }
-
-                            // For simplicity while things are in progress we keep all the endpoint statements in their own
-                            // block and now at the last moment we move them out:
-                            if (_usesMinimalHostingModel)
-                            {
-                                var lambdaParent = ((CSharpStatement)lambdaStatement.Parent);
-
-                                var statements = lambdaStatement.Statements.ToArray();
-                                var targetBlock = lambdaParent.Parent;
-                                var index = targetBlock.Statements.IndexOf(lambdaParent);
-                                targetBlock.InsertStatements(index, statements);
-                                lambdaParent.Remove();
-
-                            }
-                        }, int.MaxValue);
-                    },
-                    priority: 0);*/
-
                 foreach (var request in _serviceConfigurationRequests)
                 {
                     ProcessServiceConfigurationRequest(request);
@@ -187,7 +148,7 @@ internal class AppStartupFile : IAppStartupFile
 
                 isBuilt = true;
 
-                // Our previous subscriptions are configured to only process if "isBuilt" is false
+                // Our previous subscriptions are configured to only process if "isBuilt" is false,
                 // and we make new subscriptions below. This is because other templates which are
                 // also subscribed may have been (by chance) instantiated after this class,
                 // resulting in their handlers being registered after and thus would only get to
@@ -202,14 +163,34 @@ internal class AppStartupFile : IAppStartupFile
 
     public IAppStartupFile ConfigureServices(Action<IHasCSharpStatements, IServiceConfigurationContext> configure)
     {
-        IHasCSharpStatements block = (_usesMinimalHostingModel, _usesTopLevelStatements) switch
+        if (_configureServicesBlock == null)
         {
-            (false, false) or (false, true) => _cSharpFile.Classes.Single().Methods.First(x => x.Name == "ConfigureServices"),
-            (true, false) => _cSharpFile.Classes.Single().Methods.First(x => x.Name == "Main"),
-            (true, true) => _cSharpFile.TopLevelStatements
-        };
+            switch (_usesMinimalHostingModel, _usesTopLevelStatements)
+            {
+                case (false, false) or (false, true):
+                {
+                    _configureServicesBlock = _cSharpFile.Classes.Single().Methods.First(x => x.Name == "ConfigureServices");
+                    break;
+                }
+                case (true, false):
+                {
+                    var block = new ServiceRegistrationStatementBlock();
+                    _configureServicesBlock = block;
+                    _cSharpFile.Classes.Single().Methods.First(x => x.Name == "Main").AddStatement(block);
+                    break;
 
-        configure(block, _serviceConfigurationContext);
+                }
+                case (true, true):
+                {
+                    var block = new ServiceRegistrationStatementBlock();
+                    _configureServicesBlock = block;
+                    _cSharpFile.TopLevelStatements.AddStatement(block);
+                    break;
+                }
+            }
+        }
+
+        configure(_configureServicesBlock, _serviceConfigurationContext);
 
         return this;
     }
@@ -228,7 +209,7 @@ internal class AppStartupFile : IAppStartupFile
         return this;
     }
 
-    public IAppStartupFile ConfigureEndpoints(Action<IHasCSharpStatements, IEndpointConfigurationContext> configure)
+    public IAppStartupFile ConfigureEndpoints(Action<IHasCSharpStatements?, IEndpointConfigurationContext> configure)
     {
         ConfigureApp((statements, _) =>
         {
@@ -251,12 +232,15 @@ internal class AppStartupFile : IAppStartupFile
 
     public IAppStartupFile AddServiceConfiguration<TStatement>(
         Func<IServiceConfigurationContext, TStatement> create,
-        ConfigureServiceStatement<TStatement> configure = null,
+        ConfigureServiceStatement<TStatement>? configure = null,
         int? priority = null) where TStatement : CSharpStatement
     {
-        var statement = create(_serviceConfigurationContext);
-        Insert(statement, StatementType.ConfigureServices, priority);
-        configure?.Invoke(statement, _serviceConfigurationContext);
+        ConfigureServices((statements, context) =>
+        {
+            var statement = create(context);
+            Insert(statements, statement, StatementType.ConfigureServices, priority);
+            configure?.Invoke(statement, context);
+        });
 
         return this;
     }
@@ -264,12 +248,12 @@ internal class AppStartupFile : IAppStartupFile
     public IAppStartupFile AddServiceConfigurationLambda(
         string methodName,
         IEnumerable<string> parameters,
-        ConfigureServiceLambda configure = null,
+        ConfigureServiceLambda? configure = null,
         int? priority = null)
     {
-        ConfigureServices((statements, _) =>
+        ConfigureServices((statements, context) =>
         {
-            var invocationStatement = (CSharpInvocationStatement)statements.Statements.FirstOrDefault(x =>
+            var invocationStatement = (CSharpInvocationStatement?)statements.Statements.FirstOrDefault(x =>
                 x.TryGetMetadata<string>("lambda-registration-for", out var name) &&
                 name == methodName);
             var startupLambda = invocationStatement?.Statements.OfType<StartupLambda>().First();
@@ -278,11 +262,11 @@ internal class AppStartupFile : IAppStartupFile
             {
                 startupLambda = new StartupLambda(parameters);
 
-                invocationStatement = new CSharpInvocationStatement($"{_serviceConfigurationContext.Services}.{methodName}");
+                invocationStatement = new CSharpInvocationStatement($"{context.Services}.{methodName}");
                 invocationStatement.AddMetadata("lambda-registration-for", methodName);
                 invocationStatement.AddArgument(startupLambda);
 
-                Insert(invocationStatement, StatementType.ConfigureServices, priority);
+                Insert(statements, invocationStatement, StatementType.ConfigureServices, priority);
             }
             else
             {
@@ -293,8 +277,8 @@ internal class AppStartupFile : IAppStartupFile
                 statement: invocationStatement,
                 lambdaStatement: startupLambda,
                 context: new ServiceConfigurationLambdaContext(
-                    Configuration: _serviceConfigurationContext.Configuration,
-                    Services: _serviceConfigurationContext.Services,
+                    Configuration: context.Configuration,
+                    Services: context.Services,
                     Parameters: startupLambda.Parameters));
         });
 
@@ -303,12 +287,15 @@ internal class AppStartupFile : IAppStartupFile
 
     public IAppStartupFile AddContainerRegistration<TStatement>(
         Func<IServiceConfigurationContext, TStatement> create,
-        ConfigureServiceStatement<TStatement> configure = null,
+        ConfigureServiceStatement<TStatement>? configure = null,
         int? priority = null) where TStatement : CSharpStatement
     {
-        var statement = create(_serviceConfigurationContext);
-        Insert(statement, StatementType.ContainerRegistration, priority);
-        configure?.Invoke(statement, _serviceConfigurationContext);
+        ConfigureServices((statements, context) =>
+        {
+            var statement = create(context);
+            Insert(statements, statement, StatementType.ContainerRegistration, priority);
+            configure?.Invoke(statement, context);
+        });
 
         return this;
     }
@@ -316,13 +303,12 @@ internal class AppStartupFile : IAppStartupFile
     public IAppStartupFile AddContainerRegistrationLambda(
         string methodName,
         IEnumerable<string> parameters,
-        ConfigureServiceLambda configure = null,
+        ConfigureServiceLambda? configure = null,
         int? priority = null)
     {
-
-        ConfigureServices((statements, _) =>
+        ConfigureServices((statements, context) =>
         {
-            var invocationStatement = (CSharpInvocationStatement)statements.Statements.FirstOrDefault(x =>
+            var invocationStatement = (CSharpInvocationStatement?)statements.Statements.FirstOrDefault(x =>
                 x.TryGetMetadata<string>("lambda-registration-for", out var name) &&
                 name == methodName);
             var startupLambda = invocationStatement?.Statements.OfType<StartupLambda>().First();
@@ -331,11 +317,11 @@ internal class AppStartupFile : IAppStartupFile
             {
                 startupLambda = new StartupLambda(parameters);
 
-                invocationStatement = new CSharpInvocationStatement($"{_serviceConfigurationContext.Services}.{methodName}");
+                invocationStatement = new CSharpInvocationStatement($"{context.Services}.{methodName}");
                 invocationStatement.AddMetadata("lambda-registration-for", methodName);
                 invocationStatement.AddArgument(startupLambda);
 
-                Insert(invocationStatement, StatementType.ContainerRegistration, priority);
+                Insert(statements, invocationStatement, StatementType.ContainerRegistration, priority);
             }
             else
             {
@@ -346,8 +332,8 @@ internal class AppStartupFile : IAppStartupFile
                 statement: invocationStatement,
                 lambdaStatement: startupLambda,
                 context: new ServiceConfigurationLambdaContext(
-                    Configuration: _serviceConfigurationContext.Configuration,
-                    Services: _serviceConfigurationContext.Services,
+                    Configuration: context.Configuration,
+                    Services: context.Services,
                     Parameters: startupLambda.Parameters));
         });
 
@@ -356,12 +342,15 @@ internal class AppStartupFile : IAppStartupFile
 
     public IAppStartupFile AddAppConfiguration<TStatement>(
         Func<IAppConfigurationContext, TStatement> create,
-        ConfigureAppStatement<TStatement> configure = null,
+        ConfigureAppStatement<TStatement>? configure = null,
         int? priority = null) where TStatement : CSharpStatement
     {
-        var statement = create(_appConfigurationContext);
-        Insert(statement, StatementType.AppConfiguration, priority);
-        configure?.Invoke(statement, _appConfigurationContext);
+        ConfigureApp((statements, context) =>
+        {
+            var statement = create(context);
+            Insert(statements, statement, StatementType.AppConfiguration, priority);
+            configure?.Invoke(statement, context);
+        });
 
         return this;
     }
@@ -369,7 +358,7 @@ internal class AppStartupFile : IAppStartupFile
     public IAppStartupFile AddAppConfigurationLambda(
         string methodName,
         IEnumerable<string> parameters,
-        ConfigureAppLambda configure = null,
+        ConfigureAppLambda? configure = null,
         int? priority = null)
     {
         if (methodName == "UseEndpoints")
@@ -380,15 +369,15 @@ internal class AppStartupFile : IAppStartupFile
         return AddAppConfigurationLambdaInternal(methodName, parameters, configure, priority);
     }
 
-    private IAppStartupFile AddAppConfigurationLambdaInternal(
+    private AppStartupFile AddAppConfigurationLambdaInternal(
         string methodName,
         IEnumerable<string> parameters,
-        ConfigureAppLambda configure,
+        ConfigureAppLambda? configure,
         int? priority)
     {
         ConfigureApp((statements, _) =>
         {
-            var invocationStatement = (CSharpInvocationStatement)statements.Statements.FirstOrDefault(x =>
+            var invocationStatement = (CSharpInvocationStatement?)statements.Statements.FirstOrDefault(x =>
                 x.TryGetMetadata<string>("lambda-registration-for", out var name) &&
                 name == methodName);
             var startupLambda = invocationStatement?.Statements.OfType<StartupLambda>().First();
@@ -401,7 +390,7 @@ internal class AppStartupFile : IAppStartupFile
                 invocationStatement.AddMetadata("lambda-registration-for", methodName);
                 invocationStatement.AddArgument(startupLambda);
 
-                Insert(invocationStatement, StatementType.AppConfiguration, priority);
+                Insert(statements, invocationStatement, StatementType.AppConfiguration, priority);
             }
             else
             {
@@ -423,7 +412,7 @@ internal class AppStartupFile : IAppStartupFile
 
     public IAppStartupFile AddUseEndpointsStatement<TStatement>(
         Func<IEndpointConfigurationContext, TStatement> create,
-        ConfigureEndpointStatement<TStatement> configure = null,
+        ConfigureEndpointStatement<TStatement>? configure = null,
         int? priority = null) where TStatement : CSharpStatement
     {
         priority ??= int.MaxValue;
@@ -434,7 +423,7 @@ internal class AppStartupFile : IAppStartupFile
 
         AddAppConfigurationLambdaInternal(
             methodName: "UseEndpoints",
-            parameters: new[] { "endpoints" },
+            parameters: ["endpoints"],
             configure: (_, lambdaStatement, _) =>
             {
                 var insertBelow = lambdaStatement.Statements
@@ -461,28 +450,13 @@ internal class AppStartupFile : IAppStartupFile
         return this;
     }
 
-    private void Insert(CSharpStatement statement, StatementType type, int? priority)
+    private static void Insert(
+        IHasCSharpStatements hasCSharpStatements,
+        CSharpStatement statement,
+        StatementType type,
+        int? priority)
     {
-        IHasCSharpStatements hasCSharpStatements;
         priority ??= int.MaxValue;
-
-        if (_usesMinimalHostingModel)
-        {
-            hasCSharpStatements = _usesTopLevelStatements
-                ? _cSharpFile.TopLevelStatements
-                : _cSharpFile.Classes.First().Methods.First(x => x.Name == "Main");
-        }
-        else
-        {
-            var @class = _cSharpFile.Classes.First();
-            hasCSharpStatements = @class.Methods.Single(x => x.Name == type switch
-            {
-                StatementType.ConfigureServices => "ConfigureServices",
-                StatementType.ContainerRegistration => "ConfigureServices",
-                StatementType.AppConfiguration => "Configure",
-                _ => throw new ArgumentOutOfRangeException(nameof(type), type, null)
-            });
-        }
 
         statement
             .AddMetadata("startup-statement-type", type)
@@ -582,9 +556,9 @@ internal class AppStartupFile : IAppStartupFile
             return (request.ConcreteType.StartsWith("typeof("), request.InterfaceType != null) switch
             {
                 (false, false) => $"{ctx.Services}.{RegistrationType(request)}<{_template.UseType(request.ConcreteType)}>();",
-                (false, true) => $"{ctx.Services}.{RegistrationType(request)}<{_template.UseType(request.InterfaceType)}, {_template.UseType(request.ConcreteType)}>();",
+                (false, true) => $"{ctx.Services}.{RegistrationType(request)}<{_template.UseType(request.InterfaceType!)}, {_template.UseType(request.ConcreteType)}>();",
                 (true, false) => $"{ctx.Services}.{RegistrationType(request)}({UseTypeOf(request.ConcreteType)});",
-                (true, true) => $"{ctx.Services}.{RegistrationType(request)}({UseTypeOf(request.InterfaceType)}, {UseTypeOf(request.ConcreteType)});"
+                (true, true) => $"{ctx.Services}.{RegistrationType(request)}({UseTypeOf(request.InterfaceType!)}, {UseTypeOf(request.ConcreteType)});"
             };
 
             string UseTypeOf(string type)
@@ -629,7 +603,7 @@ internal class AppStartupFile : IAppStartupFile
                 switch (parameter)
                 {
                     case ApplicationBuilderRegistrationRequest.ParameterType.Configuration:
-                        parameterList.Add("Configuration");
+                        parameterList.Add(ctx.Configuration);
                         break;
                     case ApplicationBuilderRegistrationRequest.ParameterType.HostEnvironment:
                     case ApplicationBuilderRegistrationRequest.ParameterType.WebHostEnvironment:
@@ -686,5 +660,17 @@ internal class AppStartupFile : IAppStartupFile
         ConfigureServices,
         ContainerRegistration,
         AppConfiguration
+    }
+
+    private class ServiceRegistrationStatementBlock() : CSharpStatement(string.Empty), IHasCSharpStatements
+    {
+        public IList<CSharpStatement> Statements { get; } = new List<CSharpStatement>();
+
+        public override string GetText(string indentation)
+        {
+            return indentation + RelativeIndentation + Statements.ConcatCode($"{indentation}{RelativeIndentation}    ").TrimStart();
+        }
+
+        bool IHasCSharpStatementsActual.IsCodeBlock => true;
     }
 }