+++ Modules/Intent.Modules.AspNetCore/release-notes.md
@@ -1,3 +1,7 @@
+### Version 6.0.8
+
+- Fixed: When using minimal hosting model some service registration statements would be generated after application configuration statements.
+
 ### Version 6.0.7
 
 - Fixed: `Configuration` not resolving correctly for Application Builder registrations